moved class to the framework
